requirements.txt
bcrypt==4.2.1
bidict==0.23.1
blinker==1.9.0
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
dnspython==2.7.0
eventlet==0.39.0
Flask==3.1.0
Flask-SocketIO==5.5.1
greenlet==3.1.1
h11==0.14.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
MarkupSafe==3.0.2
pymongo==4.11.1
python-dotenv==1.0.1
python-engineio==4.11.2
python-socketio==5.12.1
requests==2.32.3
simple-websocket==1.1.0
urllib3==2.3.0
Werkzeug==3.1.3
wsproto==1.2.0
